Adds SEM class which works with all representations
